# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ScanLabelBody {
  /** Base64-encoded image of the product label */
  imageBase64: string;
}

export interface ScannedProduct {
  /** Product name extracted from the label */
  name: string;
  /**
   * Price extracted from the label
   * @nullable
   */
  price?: number | null;
  /**
   * Quantity/weight/volume extracted from the label
   * @nullable
   */
  quantity?: string | null;
  /**
   * Brand name extracted from the label
   * @nullable
   */
  brand?: string | null;
  /**
   * Any additional notes about the product
   * @nullable
   */
  notes?: string | null;
}

export interface ShoppingList {
  id: number;
  name: string;
  createdAt: string;
}

export interface ListItem {
  id: number;
  listId: number;
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  quantity?: string | null;
  units: number;
  /** @nullable */
  notes?: string | null;
  checked: boolean;
  createdAt: string;
}

export interface ShoppingListWithItems {
  id: number;
  name: string;
  createdAt: string;
  items: ListItem[];
}

export interface CreateListBody {
  name: string;
}

export interface AddItemBody {
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  units?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateItemBody {
  /** @nullable */
  checked?: boolean | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  units?: number | null;
  /** @nullable */
  notes?: string | null;
}