# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Uploads a base64 image of a product label and returns extracted product info
 * @summary Scan a product label image
 */
export const ScanLabelBody = zod.object({
  imageBase64: zod
    .string()
    .describe("Base64-encoded image of the product label"),
});

export const ScanLabelResponse = zod.object({
  name: zod.string().describe("Product name extracted from the label"),
  price: zod.number().nullish().describe("Price extracted from the label"),
  quantity: zod
    .string()
    .nullish()
    .describe("Quantity\/weight\/volume extracted from the label"),
  brand: zod.string().nullish().describe("Brand name extracted from the label"),
  notes: zod
    .string()
    .nullish()
    .describe("Any additional notes about the product"),
});

/**
 * @summary Get all shopping lists
 */
export const GetListsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  createdAt: zod.date(),
});
export const GetListsResponse = zod.array(GetListsResponseItem);

/**
 * @summary Create a new shopping list
 */
export const CreateListBody = zod.object({
  name: zod.string(),
});

/**
 * @summary Get a shopping list with its items
 */
export const GetListParams = zod.object({
  id: zod.coerce.number(),
});

export const getListResponseItemsItemUnitsDefault = 1;

export const GetListResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  createdAt: zod.date(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      listId: zod.number(),
      name: zod.string(),
      brand: zod.string().nullish(),
      price: zod.number().nullish(),
      quantity: zod.string().nullish(),
      units: zod.number().default(getListResponseItemsItemUnitsDefault),
      notes: zod.string().nullish(),
      checked: zod.boolean(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a shopping list
 */
export const DeleteListParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteListResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Add an item to a shopping list
 */
export const AddItemToListParams = zod.object({
  id: zod.coerce.number(),
});

export const AddItemToListBody = zod.object({
  name: zod.string(),
  brand: zod.string().nullish(),
  price: zod.number().nullish(),
  quantity: zod.string().nullish(),
  units: zod.number().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Update a list item (e.g. toggle checked, change quantity)
 */
export const UpdateListItemParams = zod.object({
  id: zod.coerce.number(),
  itemId: zod.coerce.number(),
});

export const UpdateListItemBody = zod.object({
  checked: zod.boolean().nullish(),
  name: zod.string().nullish(),
  price: zod.number().nullish(),
  quantity: zod.string().nullish(),
  units: zod.number().nullish(),
  notes: zod.string().nullish(),
});

export const updateListItemResponseUnitsDefault = 1;

export const UpdateListItemResponse = zod.object({
  id: zod.number(),
  listId: zod.number(),
  name: zod.string(),
  brand: zod.string().nullish(),
  price: zod.number().nullish(),
  quantity: zod.string().nullish(),
  units: zod.number().default(updateListItemResponseUnitsDefault),
  notes: zod.string().nullish(),
  checked: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Remove an item from a shopping list
 */
export const DeleteListItemParams = zod.object({
  id: zod.coerce.number(),
  itemId: zod.coerce.number(),
});

export const DeleteListItemResponse = zod.object({
  success: zod.boolean(),
});