# lib/api-zod/src/generated/types/addItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddItemBody {
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  units?: number | null;
  /** @nullable */
  notes?: string | null;
}