# lib/api-zod/src/generated/types/listItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ListItem {
  id: number;
  listId: number;
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  quantity?: string | null;
  units: number;
  /** @nullable */
  notes?: string | null;
  checked: boolean;
  createdAt: Date;
}