# lib/api-zod/src/generated/types/scanLabelBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScanLabelBody {
  /** Base64-encoded image of the product label */
  imageBase64: string;
}