# lib/api-zod/src/generated/types/scannedProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScannedProduct {
  /** Product name extracted from the label */
  name: string;
  /**
   * Price extracted from the label
   * @nullable
   */
  price?: number | null;
  /**
   * Quantity/weight/volume extracted from the label
   * @nullable
   */
  quantity?: string | null;
  /**
   * Brand name extracted from the label
   * @nullable
   */
  brand?: string | null;
  /**
   * Any additional notes about the product
   * @nullable
   */
  notes?: string | null;
}