# lib/api-zod/src/generated/types/shoppingListWithItems.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ListItem } from "./listItem";

export interface ShoppingListWithItems {
  id: number;
  name: string;
  createdAt: Date;
  items: ListItem[];
}